refactor: move `update_keying_material()` to `ConversationGuard`
We're calling the commit-producing operation from multiple places, so, to
move all logic to `ConversationGuard`, we are introducing an `inner`
version of this function.
